@@ -72,8 +72,8 @@
         "postinstall": "electron-builder install-app-deps"
     },
     "devDependencies": {
-        "electron": "^8.0.1",
-        "electron-builder": "22.3.2"
+        "electron": "^9.2.1",
+        "electron-builder": "22.8.0"
     },
     "dependencies": {}
 }
